Fix more review comments
