+++ csharp/src/Ice/OutgoingResponse.cs
@@ -69,18 +69,3 @@ public sealed class OutgoingResponse
     {
     }
 }
-
-/// <summary>
-/// Represents the status of a reply.
-/// </summary>
-public enum ReplyStatus : byte
-{
-    Ok = 0,
-    UserException = 1,
-    ObjectNotExist = 2,
-    FacetNotExist = 3,
-    OperationNotExist = 4,
-    UnknownLocalException = 5,
-    UnknownUserException = 6,
-    UnknownException = 7
-};
